refactor: 语音通话重构 + 迁移合并 + 错误码同步
- feat(voice): 重构 useVoiceChat/useVoiceWakeup，整合 InputArea/VoiceCall 语音入口
- refactor(dao): 合并 v101_business_roles/v102_mission_hash/v103_workflow_reflections 至 v100_consolidated，删除冗余迁移
- refactor(providers): 删除 realtime_client.rs，实时逻辑下沉 gateway/realtime
- feat(harness): 新增错误码与 util_fns 工具函数，扩展 agent_session/provider trait
- chore(i18n): 同步 11 语言 error 段与语音相关翻译

+++ src-tauri/crates/agent/src/session_manager.rs
@@ -61,8 +61,8 @@ impl TokenUsageBreakdown {
         }
     }
 
-    pub fn tokens_delta(&self) -> i32 {
-        self.total_tokens as i32
+    pub fn tokens_delta(&self) -> i64 {
+        self.total_tokens as i64
     }
 }
 
@@ -404,7 +404,7 @@ impl SessionManager {
                     usage.or_else(|| session.session().messages.iter().rev().find_map(|m| m.usage));
                 let tokens_delta = effective_usage
                     .as_ref()
-                    .map(|u| u.input_tokens as i32 + u.output_tokens as i32)
+                    .map(|u| u.input_tokens as i64 + u.output_tokens as i64)
                     .unwrap_or(0);
 
                 drop(conv_index);
@@ -710,7 +710,7 @@ impl SessionManager {
         // Persist updates
         if let Some(axagent_session_id) = session.axagent_session_id() {
             let tokens_delta =
-                summary.usage.input_tokens as i32 + summary.usage.output_tokens as i32;
+                summary.usage.input_tokens as i64 + summary.usage.output_tokens as i64;
             // Cost is now calculated in agent_query command and emitted via agent-done event.
             // The DB field is kept for historical records; we store 0.0 here as the
             // authoritative cost comes from the event payload.
